The original sr_post test wasn't working, it was reporting success as long as the subscriber received the post. It did not confirm that the post was valid by a successful download.
First changed test to count successful downloads, noticed many were missing.

The download_directory setting was incorrectly set to the same directory as another test, so sometimes the
download would not happen because the file was already there, so it would be skipped.  Counts were always wrong.
set *directory* correctly, and it works now.